+++ api/src/main/kotlin/dev/lyranie/floria/api/module/ClientModule.kt
@@ -56,6 +56,4 @@ open class ClientModule(
             onDisable()
         }
     }
-
-    fun shouldRender() = toggled
 }