+++ core/src/main/kotlin/dev/lyranie/floria/mixin/PlayerEntityMixin.kt
@@ -15,17 +15,20 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package dev.lyranie.floria.event.events
+package dev.lyranie.floria.mixin
 
-import dev.lyranie.floria.api.event.ClientEvent
-import io.netty.channel.Channel
-import net.minecraft.network.packet.Packet
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo
+import dev.lyranie.floria.event.EventHandler
+import dev.lyranie.floria.event.events.ClipAtLedgeEvent
+import net.minecraft.entity.player.PlayerEntity
+import org.spongepowered.asm.mixin.Mixin
+import org.spongepowered.asm.mixin.injection.At
+import org.spongepowered.asm.mixin.injection.Inject
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable
 
-class PacketEvent(val packet: Packet<*>, val type: Type, val channel: Channel?, callbackInfo: CallbackInfo) :
-    ClientEvent(callbackInfo) {
-    enum class Type {
-        INCOMING,
-        OUTGOING
+@Mixin(PlayerEntity::class)
+class PlayerEntityMixin {
+    @Inject(method = ["clipAtLedge"], at = [At("HEAD")], cancellable = true)
+    fun clipAtLedge(callbackInfo: CallbackInfoReturnable<Boolean>) {
+        EventHandler.handleEvent(ClipAtLedgeEvent(callbackInfo))
     }
 }